% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country.data.R
\name{country.data}
\alias{country.data}
\title{Get Country Data}
\usage{
country.data(
  columns = NULL,
  countries = NULL,
  list = FALSE,
  countrydata = read.csv(system.file("extdata", "countrydata.csv", package =
    "matchcountry"), na.strings = "", stringsAsFactors = FALSE)
)
}
\arguments{
\item{columns}{Columns to return (NULL returns all)}

\item{countries}{Which countries to return (list of ISO codes)}

\item{list}{Return a vector with all possible columns?}

\item{countrydata}{The countrydata table

This file provides data on countries, updated regularly. If list = TRUE,
other parameters will be overridden and it will return a list of columns.
Otherwise it will return a data frame that includes the selected countries
and columns.}
}
\description{
Get Country Data
}
\examples{
# All country data
mc <- country.data()

country.data("country", "DEU")

country.data(c("country", "region"), c("DEU", "USA", "BRA", "ZAF"))

}
