% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.mc.R
\name{download.mc}
\alias{download.mc}
\title{Download the MatchCountry tables from Github}
\usage{
download.mc(
 
    match = "https://raw.githubusercontent.com/philbastian/matchcountry/master/inst/extdata/match.csv",
 
    countrydata = "https://raw.githubusercontent.com/philbastian/matchcountry/master/inst/extdata/countrydata.csv"
)
}
\arguments{
\item{match}{Path to the master match csv file}

\item{countrydata}{Path to the master countrydata csv file

This updates the Match Country files in the library to their current versions. It will overwrite
the existing files, so this should be used with care if reproducability of older analyses is
a priority. You can also use this to make your own custom files for matching.}
}
\description{
Download the MatchCountry tables from Github
}
\examples{
download.mc()

}
