% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.country.R
\name{match.country}
\alias{match.country}
\title{Match country}
\usage{
match.country(country, output = "iso", language = "english")
}
\arguments{
\item{country}{A vector of country names or ISO 3 digit alpha codes}

\item{output}{A column of the MatchCountry table (mc) to return}

\item{language}{The language of the inputs (by default, English; NULL will match on all)
languages)}

\item{match}{A csv file to use for matching}

\item{countrydata}{A csv file to use}
}
\value{
A vector with the same length as country with the matching results
}
\description{
Match country
}
\details{
This function is designed to recognize different variations of country names and
standardize them. For example, "St. Kitts & Nevis", "Saint Kitts & Nevis" and "St Kitts and
Nevis" all refer to the same place, but a simple merge with a table would fail to match them
all. The database has a table of common alternative names. Further, the algorithm removes
extended characters that might lead to confusion (St. vs. Saint, for example). Where a match
cannot be found, NA is returned in its place.

If no language is specified, the algorithm will match on all languages, but this is less
efficient and could be prone to errors.

The default output is the iso field of the countrydata data frame; any column of the countrydata
can be used, however.
}
\examples{

match.country("United Republic of Tanzania")
match.country("Tanzania")

}
