% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mc.R
\name{read.mc}
\alias{read.mc}
\title{Reads the MatchCountry tables into memory}
\usage{
read.mc(
  match = read.csv(system.file("extdata", "match.csv", package = "matchcountry"),
    na.strings = "", stringsAsFactors = FALSE, encoding = "UTF-8"),
  countrydata = read.csv(system.file("extdata", "countrydata.csv", package =
    "matchcountry"), na.strings = "", stringsAsFactors = FALSE, encoding = "UTF-8")
)
}
\arguments{
\item{match}{Local path to the match csv file}

\item{countrydata}{Local path to the countrydata csv file

This is mostly used behind the scenes (on package loading or when update.mc is called). It loads the
csv files with country data into memory.}
}
\description{
Reads the MatchCountry tables into memory
}
\examples{
read.mc()

}
